+++ acceptance/openstack/cce/helpers.go
@@ -38,7 +38,7 @@ func CreateCluster(t *testing.T, vpcID, subnetID string) string {
 			},
 			KubernetesSvcIpRange: "10.247.0.0/16",
 		},
-	}).Extract()
+	})
 	th.AssertNoErr(t, err)
 
 	th.AssertNoErr(t, waitForClusterToActivate(client, cluster.Metadata.Id, 30*60))
@@ -76,7 +76,7 @@ func CreateTurboCluster(t *testing.T, vpcID, subnetID string, eniSubnetID string
 			},
 			KubernetesSvcIpRange: "10.247.0.0/16",
 		},
-	}).Extract()
+	})
 	th.AssertNoErr(t, err)
 
 	th.AssertNoErr(t, waitForClusterToActivate(client, cluster.Metadata.Id, 30*60))
@@ -92,7 +92,7 @@ func DeleteCluster(t *testing.T, clusterID string) {
 
 func waitForClusterToActivate(client *golangsdk.ServiceClient, id string, secs int) error {
 	return golangsdk.WaitFor(secs, func() (bool, error) {
-		cluster, err := clusters.Get(client, id).Extract()
+		cluster, err := clusters.Get(client, id)
 		if err != nil {
 			return false, err
 		}
@@ -108,7 +108,7 @@ func waitForClusterToActivate(client *golangsdk.ServiceClient, id string, secs i
 
 func waitForClusterToDelete(client *golangsdk.ServiceClient, id string, secs int) error {
 	return golangsdk.WaitFor(secs, func() (bool, error) {
-		_, err := clusters.Get(client, id).Extract()
+		_, err := clusters.Get(client, id)
 		if err != nil {
 			if _, ok := err.(golangsdk.ErrDefault404); ok {
 				return true, nil
