+++ openstack/cce/v3/clusters/update_master_ip.go
@@ -2,13 +2,13 @@ package clusters
 
 import "github.com/opentelekomcloud/gophertelekomcloud"
 
-func UpdateMasterIp(client *golangsdk.ServiceClient, id string, opts UpdateIpOpts) (r UpdateIpResult) {
+func UpdateMasterIp(client *golangsdk.ServiceClient, id string, opts UpdateIpOpts) (err error) {
 	b, err := golangsdk.BuildRequestBody(opts, "spec")
 	if err != nil {
-		return nil, err
+		return
 	}
 
-	raw, err := client.Put(client.ServiceURL("clusters", id, "mastereip"), b, nil, &golangsdk.RequestOpts{
+	_, err = client.Put(client.ServiceURL("clusters", id, "mastereip"), b, nil, &golangsdk.RequestOpts{
 		OkCodes: []int{200},
 	})
 	return
